Set keypath to false for Checkedarray elements that are not checked.
This way, when the element is unchecked and the change event is fired, we clear
the underlying model value.

Previously, it would set the keypath to the element value whenever the change
event was fired, most likely because it was expected that change events would
only be fired when radios/checkboxes were selected, and that we'd never be
deselecting those elements.