Radio binder take 2 (#149)
* Split options_field partial into separate checkboxes and radio partials.

I plan to modify the code for radio partials and I don't want to have
a bunch of logic branches in the partial.

* Now that checkbox and radio are split, clean up partials removing outdated logic.

* Add passing tests for new checkbox/radio fixtures

* Adds dobtradiogroup custom binder.

This custom binder ensures that radio groups with the include_other_option
function properly. Previously, the main radio inputs were a checkedarray, and
the 'other' input was a separate checked input bound to 'other_checked'. This
led to this gross bug: https://github.com/dobtco/screendoor-v2/issues/4557

This custom binder pushes all the logic into the bind and routine methods. It
determines whether the 'other' input was selected by a special class on the
'other' input. It handles setting and setting the 'checked', 'other_checked' and
'other_text' attributes on the model based on whether a primary or other input
was selected.

* Delete obsolete code.

* Commit dist/

* release 1.1.2

* Adds asdf config file