Add "is default hidden" helper (#150)
* Implements a helper method that determines if this field is default hidden.

Default hidden means that on a freshly loaded form with no user input, the field
is hidden. Examples include admin_only fields, and fields with conditions that
are not met by default values.

Note: some conditions are met by default values, e.g. "field is blank"

FormRenderer.ConditionChecker is currently tightly coupled to the FormRenderer
main view, so I added a third constructor parameter that can be used in settings
where the FormRenderer main view is not initialized, such as in FormBuilder.