Refactor API FormBuilder uses to determine if field should be greyed out.
Previously we exposed a new method, isHidden that was a logical duplicate of the
existing isVisible method, but allowed the field collection to be passed in. It
turns out that Adam had already refactored isVisible and the ConditionChecker to
allow fields to be passed in from the outside.

Now we can get rid of isHidden, and extract the logic we want from isVisible and
use it for both isVisible and the new method FormBuilder will use, satisfiesConditions.